update the granular access and schema separation
